General improvements to give basic functionality
We add the RstContentCreator to convert reStructureText annotations to
the appropriate docutils nodes.

We add the LabelConverter to do the Maya-esque attribute name to title
conversion. ie. floatAtttr -> "Float Attr"

We add a basic dispatch method as a way of responding to different
keyword arguments in the editorTemplate call.

We add the "add_control" method to handle those cases and create
definition_list_items with the parameter name as the term and the
parameter's annotation as the definition.

Finally, we add the begin_layout & end_layout methods which add a rubric
node with the layout's name from begin_layout and then manage the
pushing & popping of a sublist of nodes which is then inserted into a
definition_list once all the entries in the layout section have been
gathered.